Use CPython interface tables instead of creating `__special__` methods
used by Python classes, because the former are much more efficient. On
my simple benchmark (included) the speed difference is:
#### Without my change

| Method | Time | Time / List Time |
| --- | --: | --: |
| len(s) | 0.12184501 | 2.928329542 |
| 1 in s | 0.09164882 | 3.08188086266 |
| s[0] | 0.11913800 | 5.4947823314 |
| s[0] = 1 | 0.14618182 | 5.18284868977 |
| del s2[-1] | 0.11298084 | 2.5407810925 |
| s[1:-1] | 0.31020999 | 2.68454901087 |
| s[-2:-1] = s1 | 0.19746900 | 3.87809617456 |
| del s2[-1:] | 0.14702821 | 3.26322501442 |
| s + s | 0.30015993 | 1.7862811172 |
| s \* 2 | 0.30452204 | 1.54962656327 |
| s += s1 | 0.12353301 | 2.9126216321 |
#### With my change

| Method | Time | Time / List Time |
| --- | --: | --: |
| len(s) | 0.04303408 | 1.03756545932 |
| 1 in s | 0.03646994 | 1.22752842801 |
| s[0] | 0.03031492 | 1.40444473899 |
| s[0] = 1 | 0.03436899 | 1.2166845317 |
| del s2[-1] | 0.03327703 | 0.745054288063 |
| s[1:-1] | 0.21536398 | 1.76472702763 |
| s[-2:-1] = s1 | 0.07444501 | 1.46174588387 |
| del s2[-1:] | 0.06555700 | 1.45362367109 |
| s + s | 0.21622396 | 1.29087995035 |
| s \* 2 | 0.22456384 | 1.14321346384 |
| s += s1 | 0.04029489 | 0.950439203248 |

The last column is the time relative to list. The numbers are somewhat noisy.

So most methods got a 2-4x speed improvement, with 2.65x average. Most
methods are still slower than corresponing list methods, so there are
probably more improvements possible.