@andre-merzky 2016-10-22
Hey @eltoder, 

this repo has been silent for quite a while, and we thus created a fork at https://github.com/diamondman/bitarray/.   

If you are interested, please feel free to transplant your pull request to the forked repo, we would be very happy to begin the code review and merge into bitarray before pushing out a new release.  If you do not have the time to do so, we would kindly ask your permission to do the PR transfer our-self.  If you could ping back in the next couple of days, one way or the other, that would be great.

Many thanks!