@diamondman 2016-10-24
https://github.com/diamondman/bitarray/pull/7