@ilanschnell 2020-08-07
Thanks!  And sorry for the very late response.  In bitarray 1.5.0, I'm using Python's sequence, mapping and number methods.